+++ client/.eslintrc.json
@@ -16,6 +16,7 @@
     "@typescript-eslint/member-delimiter-style": "off",
     "@typescript-eslint/no-misused-promises": "off",
     "@typescript-eslint/strict-boolean-expressions": "off",
+    "@typescript-eslint/restrict-template-expressions": "off",
     "react/function-component-definition": [
       2,
       {
